Change instantiation/serialisation scheme
To avoid forcing users to sub-class Specable,
we dropped this approach in favour of an even
more minimal scheme: We just require exposing
a .spec property that represents the setup for
a given module.

This should hopefully be enough to do sane
restore of models from disk, while not imposing
too much complexity cost on users.

The additional benefit of this is that we don't
need to depend on specable any more.