Improve spline accuracy (#2)
The default spline tolerance of 1e-2 is, in my experience, too high. By fixing one tensor that was not in float64, this PR should unlock much better spline accuracies. The default is now the same as rascaline (spline accuracies of 1e-8 and a max number of spline points of 10k)