Overhaul radial/angular interface
The (hopefully) last overhaul of the radial/angular expansions.

We switch from returning packed tensors to lists of per-degree
tensors. This is slightly less efficient in some cases, but it
moves all the logic of splitting and distinguishing cases where
we have a per-degree radial basis from one basis for all degrees
into the radial expansions themselves, which are much more suited
to this job. The SphericalExpansion therefore becomes *much* more
simple, and the effort of interfacing between radial and the rest
is reduced a fair bit.

We also reorganise radial bases into the (currently very minimal)
"simple" bases, i.e., the usual polynomial suspects, and the more
involved trimmed and splined "physical" bases, anticipating the
upcoming inclusion of more complex Laplace equation-related bases.

For consistency, we apply the same convention to the angular case.