Make cutoff and max_angular arguments to SphericalExpansion (#4)
This makes it such that cutoff and max_angular are
top-level arguments to SphericalExpansion, which are
then passed on to the radial/angular expansions.

Before, we had the radial basis as the source of truth
for cutoff, and max_angular, which could even be determined
at initalisation time with some "trimming" procedure.
I decided that this is too complicated and brittle, and
it is unintuitive to specify the fundamental parameters of
the expansion in some nested dictionary, and to even
have them be changed at runtime in an opague way.

This keeps the functionality to trim, but the case of
max_angular getting overwritten is not supported by
SphericalExpansion -- it can easily be put into some
other, less generic, expansion.

This change implies some changes to how the trimming
works, because we need to privilege max_angular over
all other choices, so we can guarantee that the radial
basis doesn't do unexpected things. Notably, this is a
change in interface for spex.radial!

This change also allows us to walk back the somewhat
dubious choice to make cutoff an argument, rather
than a parameter, to the cutoff functions, and extracts
them out of spex.radial for clarity.